refactor: move default event handlers after plugin registration
